requirements.txt
numpy
tqdm
torch
datasets
transformers
pandas
matplotlib
diffusers
open_clip_torch
pillow
wandb